Time: 0 ms (100%), Space: 43.7 MB (30.23%) - LeetHub

+++ 3688-bitwise-or-of-even-numbers-in-an-array/3688-bitwise-or-of-even-numbers-in-an-array.cs
@@ -5,7 +5,7 @@ public class Solution
         var bitwiseOr = 0;
         foreach(var num in nums)
         {
-            if((int)(num & 1) == 1) continue;
+            if((num & 1) == 1) continue;
 
             bitwiseOr |= num;
         }
